Add `on_market_close` tests
